Move subcommands into separate packages.
