+++ cli/check/check.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package check
 
 // The check subcommand searches the repo for packages using the filter provided. The default
 // filter is an empty string and will return all packages.
@@ -35,6 +35,8 @@ import (
 	"github.com/google/logger"
 	"github.com/google/subcommands"
 )
+
+func init() { subcommands.Register(&checkCmd{}, "package query") }
 
 type checkCmd struct {
 	sources string
