+++ cli/install/install_test.go
@@ -1,4 +1,4 @@
-package main
+package install
 
 import (
 	"archive/tar"
