+++ cli/rmrepo/rmrepo.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package rmrepo
 
 import (
 	"context"
@@ -25,6 +25,8 @@ import (
 	"github.com/google/googet/v2/settings"
 	"github.com/google/subcommands"
 )
+
+func init() { subcommands.Register(&rmRepoCmd{}, "repository management") }
 
 type rmRepoCmd struct{}
